Extract common logic into helper class
